Reworked folder structure
