+++ src/b_helpers.c
