+++ src/checker_bonus.c
