+++ src/checker_helpers_bonus.c
