+++ src/operations_bonus.c
