+++ src/reverse_rotate_stack.c
