+++ src/rotate_stack.c
