+++ src/rotate_stack_bonus.c
