+++ src/stack_creation.c
