+++ src/swap_stack.c
