+++ src/swap_stack_bonus.c
