+++ tester/avgs
@@ -20,3 +20,5 @@ avg of 100 tests between 1 and 5
 7
 avg of 100 tests between 1 and 5
 7
+avg of 100 tests between 1 and 500
+5184
